pyproject.toml
[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[project]
name = "mkdocstrings-python-legacy"
description = "A legacy Python handler for mkdocstrings."
authors = [{name = "Timothée Mazzucotelli", email = "dev@pawamoy.fr"}]
license = "ISC"
license-files = ["LICENSE"]
readme = "README.md"
requires-python = ">=3.9"
keywords = []
dynamic = ["version"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Documentation",
    "Topic :: Software Development",
    "Topic :: Utilities",
    "Typing :: Typed",
]
dependencies = [
    "mkdocstrings>=0.28.3",
    "mkdocs-autorefs>=1.1",
    "pytkdocs>=0.14",
]

[project.urls]
Homepage = "https://mkdocstrings.github.io/python-legacy"
Documentation = "https://mkdocstrings.github.io/python-legacy"
Changelog = "https://mkdocstrings.github.io/python-legacy/changelog"
Repository = "https://github.com/mkdocstrings/python-legacy"
Issues = "https://github.com/mkdocstrings/python-legacy/issues"
Discussions = "https://github.com/mkdocstrings/python-legacy/discussions"
Gitter = "https://gitter.im/mkdocstrings/python-legacy"
Funding = "https://github.com/sponsors/pawamoy"

[tool.pdm.version]
source = "call"
getter = "scripts.get_version:get_version"

[tool.pdm.build]
# Include as much as possible in the source distribution, to help redistributors.
excludes = ["**/.pytest_cache"]
source-includes = [
    "config",
    "docs",
    "scripts",
    "share",
    "tests",
    "duties.py",
    "mkdocs.yml",
    "*.md",
    "LICENSE",
]

[tool.pdm.build.wheel-data]
# Manual pages can be included in the wheel.
# Depending on the installation tool, they will be accessible to users.
# pipx supports it, uv does not yet, see https://github.com/astral-sh/uv/issues/4731.
data = [
    {path = "share/**/*", relative-to = "."},
]

[dependency-groups]
maintain = [
    "build>=1.2",
    "git-changelog>=2.5",
    "twine>=5.1",
    "yore>=0.3.3",
]
ci = [
    "duty>=1.6",
    "griffe>=1.6",
    "ruff>=0.4",
    "pytest>=8.2",
    "pytest-cov>=5.0",
    "pytest-randomly>=3.15",
    "pytest-xdist>=3.6",
    "mypy>=1.10",
    "types-markdown>=3.6",
    "types-pyyaml>=6.0",
]
 docs = [
    "markdown-callouts>=0.4",
    "markdown-exec>=1.8",
    "mkdocs>=1.6",
    "mkdocs-coverage>=1.0",
    "mkdocs-git-revision-date-localized-plugin>=1.2",
    "mkdocs-llmstxt>=0.1",
    "mkdocs-material>=9.5",
    "mkdocs-minify-plugin>=0.8",
    "mkdocs-section-index>=0.3",
    # YORE: EOL 3.10: Remove line.
    "tomli>=2.0; python_version < '3.11'",
]

[tool.uv]
default-groups = ["maintain", "ci", "docs"]
